Refactor Test
